requirement.txt
python-dotenv
google-generativeai
langchain
langchain_community
langchain_huggingface
langchain_text_splitters
PyPDF2
groq
fastapi
uvicorn
faiss-cpu